feat: move initialize to plugins dir
