+++ packages/solutions/app-tools/src/index.ts
@@ -10,8 +10,8 @@ import {
 } from '@modern-js/utils';
 import { CliPlugin } from '@modern-js/core';
 import { getLocaleLanguage } from '@modern-js/plugin-i18n/language-detector';
+import initializePlugin from './plugins/initialize';
 import analyzePlugin from './analyze';
-import initializePlugin from './initialize';
 import { AppTools } from './types';
 import { hooks } from './hooks';
 import { i18n } from './locale';
